refactor: 공통 Wrapper 컴포넌트 파일 경로 리팩터링 (#97)
