+++ frontend/src/pages/PU/PU-003/index.tsx
@@ -1,5 +1,5 @@
 import { UserInfoMessage } from '@/constants/PU/userInfoMessage'
-import { Wrapper } from '../components/layout/Wrapper'
+import { Wrapper } from '../../../components/layout/Wrapper'
 import { UserInfoForm } from './UserInfoForm'
 
 export default function UserInfoPage() {
